Initial experimental CVE checker (port fo TingPing's script)
